services/execution-agent/requirements.txt
python-dotenv
playwright
psycopg2-binary
Pillow
minio
httpx
fastapi
uvicorn[standard]